delete old resource methods and replace them with methods in new controller
